+++ lib/bitmap-constants.js
@@ -1,5 +1,5 @@
 module.exports = {
   PIXEL_OFFSET: 10,
-  BITS_PER_PIXEL_OFFSET: 24,
+  BITS_PER_PIXEL_OFFSET: 28,
   FILE_SIZE_OFFSET: 2
 };
