+++ test/bitmap-transformer.testX.js
