+++ test/pixel-reader.testX.js
